# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/item": {
            "get": {
                "tags": [
                    "CRUD Via MongoDB"
                ],
                "operationId": "get_list",
                "responses": {
                    "200": {
                        "description": "Returns message list item"
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    },
                    "401": {
                        "description": "Error: Unauthorized. when token is not valid."
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "CRUD Via MongoDB"
                ],
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns message success create item"
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    },
                    "500": {
                        "description": "Error: Server error."
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreProjectRequest": {
                "title": "Store Project request",
                "description": "Store Project request body data",
                "required": [
                    "title",
                    "price"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "Title item",
                        "type": "string",
                        "example": "Indomie Goreng"
                    },
                    "price": {
                        "title": "price",
                        "description": "Price of item",
                        "type": "string",
                        "example": "1000"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}